계산 및 타겟 정보 작업 구조 변경
